Various fixes for the G4 ADC sampling chain
1. We were not setting the sample time correctly.  The F4 variant set
the sample time for all channels, but in the G4 world we only did it
for channel 0, which we weren't even using!

2. Set the sampling time for the aux channels (temp and voltage) to be
distinct from the current channels.  They are driven by a high
impedence source (resistor dividers) and need the extra time.  This
removes the need for G4 specific A/D constant.

3. When switching channels, always apply ADSTP first.  The datasheet
was unclear on if this was needed.  It claims that it must be asserted
"so that no conversions are active".  In our case, we know no
conversions are active, but setting ADSTP first still seems to be
required to reliably switch channels.  Previously, the combination of
sampling window and some NOPs made it happen to work.

4. Re-order some things in the sense loop so that the now longer aux
sampling time doesn't affect our overall ISR time.